requirements.txt
Flask==2.0.1
h5py==3.1.0
html5lib==1.1
ipykernel==5.5.3
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyter-http-over-ws==0.0.8
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
numpy==1.21.0
opencv-contrib-python==4.5.2.54
Pillow==8.2.0
requests==2.25.1
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-datasets==4.0.0
tensorflow-estimator==2.5.0
tensorflow-metadata==1.1.0
